Integrate HLabelData and HLabelInfo
