+++ src/otx/core/data/dataset/classification.py
@@ -7,7 +7,7 @@ from __future__ import annotations
 
 from dataclasses import dataclass
 from functools import partial
-from typing import Callable
+from typing import Callable, TYPE_CHECKING, Any
 
 import torch
 from datumaro import Image, Label
@@ -26,12 +26,166 @@ from otx.core.data.entity.classification import (
     MultilabelClsDataEntity,
 )
 
+if TYPE_CHECKING:
+    from datumaro import Label, LabelCategories
 
 @dataclass
 class HLabelInfo(LabelInfo):
-    """Meta information of hlabel classification."""
+    """The label information represents the hierarchy.
 
-    hlabel_data: HLabelData
+    All params should be kept since they're also used at the Model API side.
+
+    :param num_multiclass_heads: the number of the multiclass heads
+    :param num_multilabel_classes: the number of multilabel classes
+    :param head_to_logits_range: the logit range of each heads
+    :param num_single_label_classes: the number of single label classes
+    :param class_to_group_idx: represents the head index and label index
+    :param all_groups: represents information of all groups
+    :param label_to_idx: index of each label
+    :param empty_multiclass_head_indices: the index of head that doesn't include any label
+                                          due to the label removing
+
+    i.e.
+    Single-selection group information (Multiclass, Exclusive)
+    {
+        "Shape": ["Rigid", "Non-Rigid"],
+        "Rigid": ["Rectangle", "Triangle"],
+        "Non-Rigid": ["Circle"]
+    }
+
+    Multi-selection group information (Multilabel)
+    {
+        "Animal": ["Lion", "Panda"]
+    }
+
+    In the case above, HlabelInfo will be generated as below.
+    NOTE, If there was only one label in the multiclass group, it will be handeled as multilabel(Circle).
+
+        num_multiclass_heads: 2  (Shape, Rigid)
+        num_multilabel_classes: 3 (Circle, Lion, Panda)
+        head_to_logits_range: {'0': (0, 2), '1': (2, 4)} (Each multiclass head have 2 labels)
+        num_single_label_classes: 4 (Rigid, Non-Rigid, Rectangle, Triangle)
+        class_to_group_idx: {
+            'Non-Rigid': (0, 0), 'Rigid': (0, 1),
+            'Rectangle': (1, 0), 'Triangle': (1, 1),
+            'Circle': (2, 0), 'Lion': (2,1), 'Panda': (2,2)
+        } (head index, label index for each head)
+        all_groups: [['Non-Rigid', 'Rigid'], ['Rectangle', 'Triangle'], ['Circle'], ['Lion'], ['Panda']]
+        label_to_idx: {
+            'Rigid': 0, 'Rectangle': 1,
+            'Triangle': 2, 'Non-Rigid': 3, 'Circle': 4
+            'Lion': 5, 'Panda': 6
+        }
+        label_tree_edges: [
+            ["Rectangle", "Rigid"], ["Triangle", "Rigid"], ["Circle", "Non-Rigid"],
+        ] # NOTE, label_tree_edges format could be changed.
+        empty_multiclass_head_indices: []
+
+    All of the member variables should be considered for the Model API.
+    https://github.com/openvinotoolkit/training_extensions/blob/develop/src/otx/algorithms/classification/utils/cls_utils.py#L97
+    """
+
+    num_multiclass_heads: int
+    num_multilabel_classes: int
+    head_idx_to_logits_range: dict[str, tuple[int, int]]
+    num_single_label_classes: int
+    class_to_group_idx: dict[str, tuple[int, int]]
+    all_groups: list[list[str]]
+    label_to_idx: dict[str, int]
+    label_tree_edges: list[list[str]]
+    empty_multiclass_head_indices: list[int]
+
+    @classmethod
+    def from_dm_label_groups(cls, dm_label_categories: LabelCategories) -> HLabelData:
+        """Generate HLabelData from the Datumaro LabelCategories.
+
+        Args:
+            dm_label_categories (LabelCategories): the label categories of datumaro.
+        """
+
+        def get_exclusive_group_info(all_groups: list[Label | list[Label]]) -> dict[str, Any]:
+            """Get exclusive group information."""
+            exclusive_groups = [g for g in all_groups if len(g) > 1]
+
+            last_logits_pos = 0
+            num_single_label_classes = 0
+            head_idx_to_logits_range = {}
+            class_to_idx = {}
+
+            for i, group in enumerate(exclusive_groups):
+                head_idx_to_logits_range[str(i)] = (last_logits_pos, last_logits_pos + len(group))
+                last_logits_pos += len(group)
+                for j, c in enumerate(group):
+                    class_to_idx[c] = (i, j)
+                    num_single_label_classes += 1
+
+            return {
+                "num_multiclass_heads": len(exclusive_groups),
+                "head_idx_to_logits_range": head_idx_to_logits_range,
+                "class_to_idx": class_to_idx,
+                "num_single_label_classes": num_single_label_classes,
+            }
+
+        def get_single_label_group_info(
+            all_groups: list[Label | list[Label]],
+            num_exclusive_groups: int,
+        ) -> dict[str, Any]:
+            """Get single label group information."""
+            single_label_groups = [g for g in all_groups if len(g) == 1]
+
+            class_to_idx = {}
+
+            for i, group in enumerate(single_label_groups):
+                class_to_idx[group[0]] = (num_exclusive_groups, i)
+
+            return {
+                "num_multilabel_classes": len(single_label_groups),
+                "class_to_idx": class_to_idx,
+            }
+
+        def merge_class_to_idx(
+            exclusive_ctoi: dict[str, tuple[int, int]],
+            single_label_ctoi: dict[str, tuple[int, int]],
+        ) -> dict[str, tuple[int, int]]:
+            """Merge the class_to_idx information from exclusive and single_label groups."""
+
+            def put_key_values(src: dict, dst: dict) -> None:
+                """Put key and values from src to dst."""
+                for k, v in src.items():
+                    dst[k] = v
+
+            class_to_idx: dict[str, tuple[int, int]] = {}
+            put_key_values(exclusive_ctoi, class_to_idx)
+            put_key_values(single_label_ctoi, class_to_idx)
+            return class_to_idx
+
+        def get_label_tree_edges(dm_label_items: list[LabelCategories]) -> list[list[str]]:
+            """Get label tree edges information. Each edges represent [child, parent]."""
+            return [[item.name, item.parent] for item in dm_label_items if item.parent != ""]
+
+        all_groups = [label_group.labels for label_group in dm_label_categories.label_groups]
+
+        exclusive_group_info = get_exclusive_group_info(all_groups)
+        single_label_group_info = get_single_label_group_info(all_groups, exclusive_group_info["num_multiclass_heads"])
+
+        merged_class_to_idx = merge_class_to_idx(
+            exclusive_group_info["class_to_idx"],
+            single_label_group_info["class_to_idx"],
+        )
+
+        return HLabelInfo(
+            label_names=[item.name for item in dm_label_categories.items],
+            label_groups=all_groups,
+            num_multiclass_heads=exclusive_group_info["num_multiclass_heads"],
+            num_multilabel_classes=single_label_group_info["num_multilabel_classes"],
+            head_idx_to_logits_range=exclusive_group_info["head_idx_to_logits_range"],
+            num_single_label_classes=exclusive_group_info["num_single_label_classes"],
+            class_to_group_idx=merged_class_to_idx,
+            all_groups=all_groups,
+            label_to_idx=dm_label_categories._indices,  # noqa: SLF001
+            label_tree_edges=get_label_tree_edges(dm_label_categories.items),
+            empty_multiclass_head_indices=[],  # consider the label removing case
+        )
 
 
 class OTXMulticlassClsDataset(OTXDataset[MulticlassClsDataEntity]):
@@ -118,25 +272,34 @@ class OTXHlabelClsDataset(OTXDataset[HlabelClsDataEntity]):
         self.dm_categories = self.dm_subset.categories()[AnnotationType.label]
 
         # Hlabel classification used HLabelInfo to insert the HLabelData.
-        self.label_info = HLabelInfo(
-            label_names=[category.name for category in self.dm_categories],
-            label_groups=[label_group.labels for label_group in self.dm_categories.label_groups],
-            hlabel_data=HLabelData.from_dm_label_groups(self.dm_categories),
-        )
+        self.label_info = HLabelInfo.from_dm_label_groups(self.dm_categories)
 
-        if self.label_info.hlabel_data.num_multiclass_heads == 0:
+        if self.label_info.num_multiclass_heads == 0:
             msg = "The number of multiclass heads should be larger than 0."
             raise ValueError(msg)
+        
+        for dm_item in self.dm_subset:
+            self._add_ancestors(dm_item.annotations)
     
     def _add_ancestors(self, label_anns: list[Label]):
+        """Add ancestors recursively if some label miss the ancestor information.
+        
+        If the label tree likes below, 
+        object - vehicle -- car
+                         |- bus
+                         |- truck        
+        And annotation = ['car'], it should be ['car', 'vehicle', 'object'], to include the ancestor.
+        
+        This function add the ancestors to the annotation if missing.
+        """
         def _label_idx_to_name(idx: int) -> str:
             return self.label_info.label_names[idx]
         
         def _label_name_to_idx(name: str) -> int:
-            return [idx for idx, val in enumerate(self.label_info.label_names) if val == name]
+            return [idx for idx, val in enumerate(self.label_info.label_names) if val == name][0]
         
         def _get_label_group_idx(label_name: str) -> str:
-            return self.label_info.hlabel_data.class_to_group_idx[label_name][0]
+            return self.label_info.class_to_group_idx[label_name][0]
         
         def _find_ancestor_recursively(label_name: str, ancestors: list) -> None:
             _, dm_label_category = self.dm_categories.find(label_name)
@@ -152,8 +315,6 @@ class OTXHlabelClsDataset(OTXDataset[HlabelClsDataEntity]):
         
         all_label_names = _get_all_label_names_in_anns(label_anns)
         ancestor_dm_labels = []
-        print()
-        print("label info: ", self.label_info)
         for ann in label_anns:
             label_idx = ann.label
             label_name = _label_idx_to_name(label_idx)
@@ -161,18 +322,14 @@ class OTXHlabelClsDataset(OTXDataset[HlabelClsDataEntity]):
             
             for i, ancestor in enumerate(ancestors):
                 if ancestor not in all_label_names:
-                    ancestor_dm_labels.extend(
+                    ancestor_dm_labels.append(
                         Label(
                             label=_label_name_to_idx(ancestor),
                             id=len(label_anns)+i,
                             group=_get_label_group_idx(ancestor)
                         )
                     )
-        print()
-        print("label_anns: ", label_anns)
-        print("ancestors: ", ancestor_dm_labels)
-        breakpoint()
-        label_anns.extend(ancestor_dm_labels) 
+        label_anns.extend(ancestor_dm_labels)
 
     def _get_item_impl(self, index: int) -> HlabelClsDataEntity | None:
         item = self.dm_subset.get(id=self.ids[index], subset=self.dm_subset.name)
@@ -222,8 +379,8 @@ class OTXHlabelClsDataset(OTXDataset[HlabelClsDataEntity]):
             msg = f"The type of label_info should be HLabelInfo, got {type(self.label_info)}."
             raise TypeError(msg)
 
-        num_multiclass_heads = self.label_info.hlabel_data.num_multiclass_heads
-        num_multilabel_classes = self.label_info.hlabel_data.num_multilabel_classes
+        num_multiclass_heads = self.label_info.num_multiclass_heads
+        num_multilabel_classes = self.label_info.num_multilabel_classes
 
         class_indices = [0] * (num_multiclass_heads + num_multilabel_classes)
         for i in range(num_multiclass_heads):
@@ -231,7 +388,7 @@ class OTXHlabelClsDataset(OTXDataset[HlabelClsDataEntity]):
 
         for ann in label_anns:
             ann_name = self.dm_categories.items[ann.label].name
-            group_idx, in_group_idx = self.label_info.hlabel_data.class_to_group_idx[ann_name]
+            group_idx, in_group_idx = self.label_info.class_to_group_idx[ann_name]
 
             if group_idx < num_multiclass_heads:
                 class_indices[group_idx] = in_group_idx
