+++ src/otx/core/model/entity/classification.py
@@ -475,16 +475,16 @@ class OTXHlabelClsModel(
 
         label_info: HLabelInfo = self.label_info  # type: ignore[assignment]
         hierarchical_config["cls_heads_info"] = {
-            "num_multiclass_heads": label_info.hlabel_data.num_multiclass_heads,
-            "num_multilabel_classes": label_info.hlabel_data.num_multilabel_classes,
-            "head_idx_to_logits_range": label_info.hlabel_data.head_idx_to_logits_range,
-            "num_single_label_classes": label_info.hlabel_data.num_single_label_classes,
-            "class_to_group_idx": label_info.hlabel_data.class_to_group_idx,
-            "all_groups": label_info.hlabel_data.all_groups,
-            "label_to_idx": label_info.hlabel_data.label_to_idx,
-            "empty_multiclass_head_indices": label_info.hlabel_data.empty_multiclass_head_indices,
+            "num_multiclass_heads": label_info.num_multiclass_heads,
+            "num_multilabel_classes": label_info.num_multilabel_classes,
+            "head_idx_to_logits_range": label_info.head_idx_to_logits_range,
+            "num_single_label_classes": label_info.num_single_label_classes,
+            "class_to_group_idx": label_info.class_to_group_idx,
+            "all_groups": label_info.all_groups,
+            "label_to_idx": label_info.label_to_idx,
+            "empty_multiclass_head_indices": label_info.empty_multiclass_head_indices,
         }
-        hierarchical_config["label_tree_edges"] = label_info.hlabel_data.label_tree_edges
+        hierarchical_config["label_tree_edges"] = label_info.label_tree_edges
 
         parameters["metadata"].update(
             {
@@ -521,13 +521,13 @@ class MMPretrainHlabelClsModel(OTXHlabelClsModel):
         self.classification_layers = classification_layers
         return model
 
-    def set_hlabel_data(self, hierarchical_info: HLabelData) -> None:
+    def set(self, hierarchical_info: HLabelData) -> None:
         """Set hierarchical information in model head.
 
         Args:
             hierarchical_info: the label information represents the hierarchy.
         """
-        self.model.head.set_hlabel_data(hierarchical_info)
+        self.model.head.set(hierarchical_info)
 
     def _customize_inputs(self, entity: HlabelClsBatchDataEntity) -> dict[str, Any]:
         from mmpretrain.structures import DataSample
@@ -728,7 +728,7 @@ class OVHlabelClassificationModel(
             model_api_configuration,
         )
 
-    def set_hlabel_data(self, hierarchical_info: HLabelData) -> None:
+    def set(self, hierarchical_info: HLabelData) -> None:
         """Set hierarchical information in model head.
 
         Since OV IR model consist of all required hierarchy information,
