+++ src/otx/core/model/module/classification.py
@@ -238,7 +238,6 @@ class OTXHlabelClsLitModule(OTXLitModule):
             scheduler=scheduler,
             metric=metric,
         )
-        self.hlabel_data: HLabelData
 
     def configure_metric(self) -> None:
         """Configure the metric."""
@@ -249,7 +248,7 @@ class OTXHlabelClsLitModule(OTXLitModule):
                 if name in ["num_multiclass_heads", "num_multilabel_classes"]:
                     param_dict[name] = getattr(self.model, name)
                 elif name == "head_logits_info":
-                    param_dict[name] = self.hlabel_data.head_idx_to_logits_range
+                    param_dict[name] = self.head_idx_to_logits_range
                 else:
                     param_dict[name] = param.default
             param_dict.pop("kwargs", {})
@@ -270,15 +269,13 @@ class OTXHlabelClsLitModule(OTXLitModule):
             msg = f"The type of self.label_info should be HLabelInfo, got {type(self.label_info)}."
             raise TypeError(msg)
 
-        self.hlabel_data = self.label_info.hlabel_data
-
         # Set the OTXHlabelClsModel params to make proper hlabel setup.
-        self.model.set_hlabel_data(self.hlabel_data)
+        self.model.set_hlabel_info(self.label_info)
 
         # Set the OTXHlabelClsLitModule params.
         self.num_labels = len(self.label_info.label_names)
-        self.num_multiclass_heads = self.hlabel_data.num_multiclass_heads
-        self.num_multilabel_classes = self.hlabel_data.num_multilabel_classes
+        self.num_multiclass_heads = self.num_multiclass_heads
+        self.num_multilabel_classes = self.num_multilabel_classes
         self.num_singlelabel_classes = self.num_labels - self.num_multilabel_classes
 
     def _log_metrics(self, meter: Metric, key: str) -> None:
